requirements.txt
# Core dependencies
openai>=1.12.0
anthropic>=0.21.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0

# Data and storage
kaggle>=1.6.0
sqlalchemy>=2.0.0

# Evaluation and monitoring
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.7.0

